refactored user password encryption & verification
This has all been tested with a local SQL database recreation and all works as intented.

I changed the User controller & converter, the token service & tests and the context data accordingly to use the Microsoft.aspNetCore.Identity Library PasswordHasher methods in the Service layer for encryption and verification of passwords.

This is one of the refactorings to process feedback received during our code assessment.